Add WaitForExternalEvent with CancellationToken overloads (#820)
Previously, `WaitForExternalEvent` when invoked with a specified
timeout value would create a Timer instance that was uncancelable by
the surrounding orchestration.  This could have the practical effect of
making orchestrations hang until either the event was received (causing
the timer to be cancelled) or the timeout was reached (because the
timer expired).

This change adds overloads for `TimeSpan`-accepting overloads of
`WaitForExternalEvent` that allow providing a `CancellationToken`,
which the calling orchestration can cancel if it no longer intends to
wait for the event to be raised.