Utilize Dependency Injection for storage providers (#974)
Currently we have the following dependencies in our v2 branch.

1. DurableTask.AzureStorage
2. DurableTask.Redis
3. DurableTask.Emulator

Requiring the customers to explicitly take on these dependencies, even if they do not use them, is not a sustainable approach to adding new storage providers. It also forces us to explicitly add support for storage providers in order for customers to use it.

This PR strips out the explicit Redis and Emulator dependencies for customers, and adds support for using Dependency Injection to bring in new backend storage providers for their application. It also provides some helper classes and a new programming model for this, which abstracts away some of the discrepancies between DurableTask.Core and Durable Functions.