Sanitize the default task hub names for Azure Storage (#1010)
In a previous release, we started using the site name as the default
task hub value. However, it is possible for the site name not to be a
valid task hub name for Azure Storage, due to having '-' characters and
not meeting the length requirements. We now ensure that no matter the
sitename, the customer will have a valid task hub.

Addresses issue #976