Exclude x-functions-key header from async polling request in CallHttp() (#1217)
In the case where CallHTTP() is called on a Durable orchestration
implementing the server-side asynchronous HTTP pattern, the
x-functions-key used to trigger the orchestration does not have
sufficient permissions to poll the status of that orchestration. Our
solution is to strip this header when copying headers from the initial
request to the polling request.