Abort Durable Task session on Functions runtime failures
When the Functions JobHost is shutting down, there are often exceptions
occuring in the function execution pipeline unrelated to customer code.
This can lead to cases where the orchestration fails unecessarily, or in
some cases where the exception is ignored, can lead to orchestrations
stuck in an eternally pending state.

We now utilize common logic to try to differentiate between actual user
code failures and failures from the functions pipeline. In the case
where it is a failure in the functions pipeline, we abort the session so
that Durable Task will try again later.

Resolves #1081