Downgrade Functions V1 .NET version to net461 (#1257)
At one point, we upgraded the full framework target framework from
net461 to net471 to reduce package size for our .NET Standard packages,
as well as to utilize some new language features. This has caused two
problems.

1. When Functions V1 customers target net462, they pull in the
netstandard2.0 version of the dll, which is meant for Functions V2/V3
customers.

2. Using any 4.7 or greater release for Functions V1 breaks when using
Microsoft.NET.Sdk.Functions 1.0.30 or greater, due to some conditional
logic relying on #if NET46. Unfortunately, this version of that package
is required to use Microsoft.Azure.WebJobs 2.3.0 instead of 2.2.0.

This PR reverts to net461 to avoid these issues, as well as some
required code changes to make it compatible with this older version.