Make worker ids unique when running in App Service VMSS environment (#1360)
Worker ids are used to generate unique lease ids for partition
management. Currently this uses Environment.MachineName, which is
currently not unique when running in Azure App Service VMSS
environments.

The App Service on VMSS team is working to make these more unique, but
this will only help within a stamp. Environment.MachineName is
guaranteed to be non-unique across stamps, so we should use a safer
value based on environment varibles present in that environment.

Addresses #1323