Fix duplicate task hub names in tests (#1366)
CIs runs tests heavily in parallel, meaning duplicate task hub names on
the storage account will cause tests to conflict with each other. This
is causing the CI to hang indefinitely until the hour time limit
expires. This is odd, as most of the tests have built in timeouts, but
it is best practice to fix the duplicate task hubs anyways.