Add configurable option for partition management strategies
DurableTask.AzureStorage now has support for an alternative partition
management strategy that should reduce split brain, but due to the lack
of lease stealing, may be a bit slower for scale out.

Currently default to use legacy partition management, and we will
default to the new approach once it is hardened.