Fix unintentional breaking change with IDurableClient.RaiseEventAsync (#1501)
While refactoring IDurableOrchestrationClient in #1422, the eventData
for an overload of IDurableOrchestrationClient.RaiseEventAsync() was
accidentally nullable. C# bind IDurableOrchestrationClient.RaiseEventAsync(string, string,
string) to IDurableOrchestrationClient.RaiseEventAsync(string, string,
string, object= null, string = null) instead of IDurableOrchestrationClient.RaiseEventAsync(string,
string, object) as was the case in previous versions of the interface.
This PR reverts that unintentional change, and adds a test to ensure
this won't happen again in the future.