Expand long running timer functionality to WaitForExternalEvent
Internally WaitForExternalEvent with a timeout uses a timer. In v2.3.0
we added support for long-running timers, we just didn't remove the
now unecessary validation from the WaitForExternalEvent API. This PR
validates the scenario with a test and removes this validation.

Note that initially this PR attempted to expand this functionality to
RetryOptions in CallActivity/SubOrchestrationWithRetry, but
unfortunately the timer used by that retry is lower level than the
Durable Functions layer, meaning we can't use our long-running support.

Addresses #1538