Make E2E tests more robust in CI pipeline (#1385) (#1553)
Our end-to-end tests currently rely on having unique taskhub names so
that tests can be run safely in parallel. This has been proven to be
hard to catch/maintain this restriction, so we add a random 4
alphanumeric suffix to each task hub to make sure that there are never
duplicate taskhub names.

The other benefit of creating non-deterministic taskhubs for each test
run is that we should no longer end up with the corrupted state problem,
where a previous test run somehow corrupted a task hub and the test now
fails until we use a new storage account or preform manual intervention.

Finally, this test adds a dummy test that deletes up to 2000 taskhubs
older than 3 days, to ensure that we don't collect to many entities in
our storage account over time.