Fix null reference in IDurableClient injected from factory (#1644)
For customers using our IDurableClientFactory introduced in v2.4.0,
using the method SignalEntityAsync results in a null reference exception
due to a lack of checking whether the client is "external".

This PR puts that check in place, but there is likely additional long
term work for the case where a customer is using IDurableClientFactory
for non-client application to get access to an IDurableClient. In that
case, the client created should not be treated as "external" and we
should do the proper checks that the expected functions exists on the
application.