Ensure using latest webhook url for HTTP management APIs
Previously, we store the value of the webhook URL during extension
initialization with the assumption that this webhook URL never changes
after startup. This stored value is used for various IDurableClient
methods, along with the out-of-proc clients when local RPC endpoints are
disabled.

However, it turns out that in the case of slot-swap operations,
WEBSITE_HOSTNAME uses the staging slot hostname at site startup, and
only after the slot swap does the hostname change. The function host
listens for a new header value sent as a part of some ping, updating the
source of ExtensionConfigContext.GetWebhookUrl() with the new value once
the slot swap is complete. This PR makes sure we always fetch the value
from the ExtensionConfigContext instead of caching it to ensure we have
the most up-to-date value.