Update release template with learnings from v2.5.0
In the v2.5.0 release, we discovered that releasing new versions of
DTFx.Redis and DTFx.Emulator is unnecessary, and leads to potential
customer confusion.

In addition, the second merge from dev into main after resetting the
release notes and the pending docs felt unecessary, as having the
release tag with those files as they were is acceptable and reduces
unnecessary work.