Adding a new base to RetryOptions: SerializableRetryOptions
This class *is* JSON Serializable and contained as the content for a
DurableHttpRequest. However the previous RetryOptions - with the Handle - are
available for use as they were before. When needed, we now create the
RetryOptions instance from the Serialiable one (with a blank Handle property)
and use that for Retrying when serialization doesn't take place.

For #1559