Fix NullReferenceExceptions in DurableClient (#1971)
Fixes a bug reported by @scale-tone (#1958).

The problem is that DurableClient objects are sometimes constructed without a reference to a DurabeExtension. This can cause NullReferenceExceptions when methods are accessing this.config. There are a number of those situations (beyond just the one found in #1958).

This PR fixes all of them:

- CleanEntityStorageAsync now gets the necessary information from the durableTaskOptions and from the durabilityProvider
- Added a constructor that can pass in a client factory, when constructing DurableClient without a DurableTaskExtension
- This client factory is used (in place of this.config.GetClient()) when constructing another client from within DurableClient
- Added null checks in a few more places