Log trace events whenever entity states are created or deleted (#1639)
Prior to this PR, there were no trace events related specifically to the creation and deletion of entity states.

I have added two trace events to clarify when entity states are created and deleted:

EntityStateCreated
EntityStateDeleted

These events include the name and id of the operation if that information is applicable, i.e. if the creation or deletion can be attributed to an individual operation. If batch processing is used (e.g. for out-of-process execution), then the operation name and id are blank.

Resolves #1108