Update release instructions to include changes to bundles and templates PRs
With the latest changes to bundles, we no longer need to create a PR to update them at every release. The bundles repo now tracks each major version, so we only need to create PRs against them when releasing a new major version, which is seldom.

As for the templates repo, now we only need to update dotnet samples (`template.json files). When creating PRs against the dotnet templates, we need to target both the v3.x and v4.x branch, so two PRs will be needed in total.